Changes to the multithreaded logging so that the data structures that are used to store sensor data are now synchronized (blocking) and volatile (so cached memory is always the most recent) with the logger thread. This should fix a bug where the sensor axis are inverted, but sometimes the logger thread read the data before the inversion was applied. The logger now only logs data when new data is available.
An option for the user to select the sensor frequency has been added.
Sensor frequency is now logged and displayed on the screen.

The plotter is now run in a background thread which allows for many more
verticies to be plotted on the screen than before... it also plots at a
much faster rate... 100Hz vs 20Hz.

Localization of the logs decimal format has been added.